Countries

+++ iOS/G3MiOSSDK/Commons/GEO/GEO2DPolygonData.cpp
@@ -39,8 +39,8 @@ GEO2DPolygonData::~GEO2DPolygonData() {
 long long GEO2DPolygonData::getCoordinatesCount() const {
   long long result = GEO2DCoordinatesData::getCoordinatesCount();
   if (_holesCoordinatesArray != NULL) {
-    const int holesCoordinatesArraySize = _holesCoordinatesArray->size();
-    for (int j = 0; j < holesCoordinatesArraySize; j++) {
+    const size_t holesCoordinatesArraySize = _holesCoordinatesArray->size();
+    for (size_t j = 0; j < holesCoordinatesArraySize; j++) {
       const std::vector<Geodetic2D*>* holeCoordinates = _holesCoordinatesArray->at(j);
       result += holeCoordinates->size();
     }
